backend/api/restaurant-search/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
httpx==0.28.1
python-dotenv==1.1.1
firebase-admin==6.4.0